requirements.txt
CosmoTech_Acceleration_Library==0.2.18
azure-identity==1.12.0
